+++ src/acceptance-tests/turbulence/kill_vm_test.go
@@ -57,8 +57,10 @@ var _ = Describe("KillVm", func() {
 		})
 
 		By("deleting the deployment", func() {
-			err := client.DeleteDeployment(etcdManifest.Name)
-			Expect(err).NotTo(HaveOccurred())
+			if !CurrentGinkgoTestDescription().Failed {
+				err := client.DeleteDeployment(etcdManifest.Name)
+				Expect(err).NotTo(HaveOccurred())
+			}
 		})
 	})
